maint/jenkins: untar MPICH tarball on local disk
The untar process of MPICH tarball is performed on local disk instead of
NFS home directory. It reduces the stress on NFS when nightly jobs
starts and improves the performance.

After the test is finished, the script will create a fake directory
structure for the file generated during the test.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>