CH4:  Add support for OFI's FI_ASYNC_IOV
Added support for an OFI feature (FI_ASYNC_IOV mode bit).
This mode bit allows MPICH to manage the storage for iovec arrays
which can be easily generated from the mpich datatypes.
This should lead to better cache locality (OFI doesn't need to track
a separate allocation).

This patch moves the iovec storage for all (except CAS) of the iovec
use cases off the stack and into allocated object storage
(for example, MPI request).

For RMA, support for a single iovec per rma has been hardcoded, with
a future patch to add support for multiple iovec arrays.

The allocation of a single iovec element is still being reused by MPICH,
possibly before the rma operation has completed, it's been moved into the
request object for this patch.  This is still erroneous
use of OFI, but we are seeking an OFI clarification as to whether
single iovec can be treated as a single buffer and whether storage allocation
is required in this case.  The allocation of a single element CAS operation
is still on the stack, pending this clarification.