Fix unused variable and label warnings
There are quite a few new warnings introduced in CH4 for unused
variables and labels. We're not going to go back and fix all of them in
MPICH right now, but at least we shouldn't introduce new ones.

This also includes removing a bunch of #if 0 code, mostly to do with
fastboxes.

Change-Id: I04f7f2f26ec28fdd9aa6d16e5052062f49ccfe96
Signed-off-by: Charles J Archer <charles.j.archer@intel.com>