CH4 OFI Netmod: handle receive cancel race condition
In at least one test ("pt2pt/issendselfcancel") a posted receive
may complete normally before the receive can be canceled. When this
happens the "cancel bit" of the request will not be set and the
"while(!canceled)" progress loop will never exit.

The fix is to check if the request is complete in the progress loop
and also check before completing the request to avoid a double
completion assertion in the mpich request code.

Signed-off-by: Charles J Archer <charles.j.archer@intel.com>