Add collectives to the ADI
Step one in removing the function pointers for all of the collective
functions is to add all of the collective calls to the ADI. When these
functions are called, they will just call back up to the MPIR functions
in the MPI layer.

This gives the device a simple way to override collectives. Just change
the MPID_collective function to call an internal device function instead
of calling back up to the MPIR_*_impl version.