CH4: modify AM protocol on window offset
This patch changes the AM protocol for RMA, in terms of how the target
address is calculated and transferred over the wire.

Before: the origin calculates the full virtual address in the target
memory space, then transfer it over the wire. The target side simply
accesses this address.
After: the origin only transfers the offset from the window base,
then the target side adds the base address to it to obtain full
virtual address.

This will eliminate the need for the origin side to maintain the base
addresses of all other ranks (MPIDI_CH4R_WINFO(win, target)->base_addr),
thus enable memory usage reduction.