CH4: correct usage of MPIR_cc_decr
In some cases interpretation of value returned by MPIR_cc_decr was
not correct. It returns (to the second argument) an boolean value
indicating whether the request is incompleted (true if incompleted).
It does *not* return a counter, as some code is wrongly assuming.

Accidentally this misinterpretation has not caused any bugs,
because the condition (c == 0) is equivalent to (!c).
However this has been really confusing.