CH4R: Implement rendezvous send/receive
This patch implements rendezvous protocol in CH4R.

It works on top of today's AM API, and I believe we can easily
extend it to adopt the new API that Pavan is proposing.

Basic protocol:
1. if msg_sz > threshold the sender sends LONG_REQ (=RTS)
2. upon receive the receiver looks up the posted queue and if a
   matching recv is posted it sends LONG_ACK (=CTS) back, along with
   a pointer value of the receive request. Otherwise put a request
   into unexpected queue, with a special flag to indicate that it's
   a LMT request.
3. when a matching receive is posted (at irecv), send LONG_ACK
   if not done yet.
4. the sender side receives LONG_ACK it sends back LONG_LMT.
5. on the receiver side upon receiving LONG_LMT, it acquires
   the corresponding request object from the packet header
   and starts the data transfer

Notes:

As far as I notice SSEND is not supported yet (=everything is sent eagerly in SSEND).