CH4: increment global counter for active message rdma read operation
Without this change the fi_read() call will increment the fi_cntr on
completion, however the global counter that tracks the number of issued
operations will not be updated. This will trip an assert in the
"win progress fence" function because the "completed count" is greater
than the "issued count".