Move info_table for windows to shared_table
In ch4 active message for one-sided communication,
the info table is only required for shared windows.
Rename the structure and change, so the struct is only
allocated and used for shared windows. Handle this in ch4 and ofi.

Note: The code for shared windows in ch4 and ofi is more or less the same.
Maybe we can just delete it ich OFI