CH4/OFI: Remove source from match bits
The OFI match bits have limited space (64 bits) which has previously
been divided:

/*
 * 0123 4567 01234567 0123 4567 01234567 0123 4567 01234567 01234567 01234567
 *     |                  |                  |
 * ^   |    context id    |       source     |       message tag
 * |   |                  |                  |
 * +---- protocol
 */

However, the source bits are not really needed for matching because the
information is already captured in the source/destination address which
is included in all of the communication calls. The only reason we were
transfering the source information was to simplify things on the
receiving end to make it easier to find out the rank of the sender.

Rather than encoding this information in the match bits, it can be
included in an extra, small data field which is available after
converting fi_tsend/fi_tinject to fi_tsenddata/fi_tinjectdata,
respectively. On the receiving end, this data field can be easily read
to get the original sender rank to populate the status object's `SOURCE`
field.

Fixes csr/mpich-opa#35
Fixes csr/mpich-opa#281