CH4/OFI: Remove unnecessary assertion
The field MPIDI_CH4_NMI_OFI_AM_CONTEXT_ID_BITS is set to 24 and the
context_id field in the MPID_Comm is uint16_t. This means that the
assertion in am_impl.h will always be true. There is no need for it and
it generates a lot of compiler warnings with certain compilers.