CH4: Remove netmod_api.h macros
The macros for the CH4 API were created to make it easy to iterate on
the API and if changes were necessary, they would be easy to propagate
everywhere. These conveniences are no longer as necessary since the API
is now pretty stable. This means we can now remove these macros and make
it easier to find our way around the code with tools like ctags.

See csr/mpich-opa#297