CH4U: Remove src from match bits and expand tag
While the netmod may choose to expose more than 2^28 bits to the user
for the user tag, CH4U overwrote this with its own value. Check to make
sure the CH4U value is actually lower than the netmod value before
overwriting it.

Remove the source rank from the match bits and move it into the message
header.