CH4: Compile out symmetric heap if unavailable
Check for the presence of the thread safe random number generator and
compile its usagt gre out of its not available. This means that on such
systems, there will be no symmetric heap. This is more of a performance
problem, but on such systems, we don't expect such a performance hit to
be as big of a concern.

Fixes csr/mpich-opa#310