CH4/OFI: Add back ability to not use immediate data
Some OFI providers do not support using the immediate data field which
means with the new scheme of putting the source rank into the data field
instead of the match bits, it might be possible that such providers
would break.

Add a new path back into the code to allow either method to work. This
involves adding a wrapper function for sending messages to avoid lots of
extra macros.