Add a test case for Rget+unlock synchronization
This patch adds a test case to make sure the synchronization in unlock
works correctly.

Essentially this program does the following:
  lock_all
  req=rget(buf)
  unlock_all
  re-use buf
  wait(req)
This program is valid but if unlock_all does not implement the synchronization
semantics correctly reusing the buffer would race with outstanding rgets.

This is a test case for csr/mpich-opa#294.